Update starter notebook (#325)
* Update starter notebook

* Try increasing precision

+++ tests/test_math.py
@@ -18,7 +18,7 @@ from .conftest import create_input
 @pytest.mark.parametrize("shape", [[3, 3], [4, 6], [10, 8, 4],])
 def test_fft2(shape):
     shape = shape + [2]
-    x = create_input(shape)
+    x = create_input(shape).double()
     out_torch = fastmri.fft2c(x).numpy()
     out_torch = out_torch[..., 0] + 1j * out_torch[..., 1]
 
@@ -33,7 +33,7 @@ def test_fft2(shape):
 @pytest.mark.parametrize("shape", [[3, 3], [4, 6], [10, 8, 4],])
 def test_ifft2(shape):
     shape = shape + [2]
-    x = create_input(shape)
+    x = create_input(shape).double()
     out_torch = fastmri.ifft2c(x).numpy()
     out_torch = out_torch[..., 0] + 1j * out_torch[..., 1]
